% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{bradley_terry}
\alias{bradley_terry}
\title{Creates a Bradley-Terry Prediction Model}
\usage{
bradley_terry(
  gameIds,
  homeTeamIds,
  awayTeamIds,
  homeScores,
  awayScores,
  isNeutralSite = FALSE
)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}
}
\value{
named list with team and home field strength values and functions to predict matchups based on team identifiers or relative strengths
}
\description{
This package creates a simple prediction model based on team identifiers and previous game results.
}
\examples{

}
