% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradleyterry.multid.R
\name{bradleyterry.multid}
\alias{bradleyterry.multid}
\alias{bradleyterry,}
\alias{ANY-method}
\title{Bradley-Terry of multiple ids}
\usage{
bradleyterry.multid(a, b, id, lambda, dataset)
}
\arguments{
\item{a}{an a value}

\item{b}{a b value}

\item{id}{a vector of ids}

\item{lambda}{a data frame with document ids and corresponding lambda values. The columns must read 'DocId' and 'Lambda'}

\item{dataset}{a dataframe with three columns of data. The columns must read 'DocIDi', 'DocIDj', and 'Choose'}
}
\value{
a dataframe of updated lambda values
}
\description{
Finds the probability of a student getting certain exam questions correctly given a parameter theta.
}
\note{
This function produces vectors for use in finding the likelihood function.
}
\examples{

toydata<-data.frame(sort(rep(seq(1,10),10)), sample(1:10, 100, replace=T), sample(c(0,1), 100, replace=T))
colnames(toydata)<-c("DocIDi", "DocIDj", "Choose")
toydata$delete<- (toydata$DocIDi==toydata$DocIDj)
toydata2<-subset(toydata,delete=="FALSE")
toydata2<-toydata2[,-4]
lambda<-data.frame(c(1:10),runif(10))
colnames(lambda)<-c('DocId', 'Lambda')
id<-seq(1:10)

bradleyterry.multid(1,1,id,lambda,toydata2)

}
\author{
Benjamin Schneider, Zoe Ang, and Hyun Woo Lim
}
