% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleModel.R
\name{ensemble_model}
\alias{ensemble_model}
\title{Creates an Ensemble Prediction Model}
\usage{
ensemble_model(
  gameIds,
  homeTeamIds,
  awayTeamIds,
  homeScores,
  awayScores,
  isNeutralSite
)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}
}
\value{
named base prediction models and ensemble prediction function
}
\description{
This package creates a complex prediction model based on team identifiers and previous game results. This prediction model is build on a foundation of many other prediction models.
}
\examples{

}
