% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameScoresStdDev.R
\name{game_scores_std_dev}
\alias{game_scores_std_dev}
\title{Creates a Game Scores Standard Deviation Prediction Model}
\usage{
game_scores_std_dev(
  gameIds,
  homeTeamIds,
  awayTeamIds,
  homeScores,
  awayScores,
  isNeutralSite = FALSE
)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}

\item{isNeutralSite}{defaults to FALSE}
}
\value{
named list with team score averages and model coefficients plus functions to predict matchups based on team identifiers or relative strengths
}
\description{
This package creates a simple prediction model based on team identifiers and previous game results.
}
\examples{

}
