% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.bt.R
\name{iterative.bt}
\alias{iterative.bt}
\alias{bradleyterry,}
\alias{ANY-method}
\title{Bradley-Terry Updated Iterations}
\usage{
iterative.bt(a, b, id, lambda, dataset, iterations)
}
\arguments{
\item{a}{an a value}

\item{b}{a b value}

\item{id}{a vector of the document id numbers that need to be updated}

\item{lambda}{a data frame with document ids and corresponding lambda values. The columns must read 'DocId' and 'Lambda'}

\item{dataset}{a dataframe with three columns of data. The columns must read 'DocIDi', 'DocIDj', and 'Choose'.}

\item{iterations}{a number of the times you want lambda values updated}
}
\value{
a data frame of document ids and their corresponding updated lambdas
}
\description{
Finds the probability of a student getting certain exam questions correctly given a parameter theta.
}
\note{
This function produces vectors for use in finding the likelihood function.
}
\examples{

toydata<-data.frame(sort(rep(seq(1,10),10)), sample(1:10, 100, replace=T), sample(c(0,1), 100, replace=T))
colnames(toydata)<-c("DocIDi", "DocIDj", "Choose")
toydata$delete<- (toydata$DocIDi==toydata$DocIDj)
toydata2<-subset(toydata,delete=="FALSE")
toydata2<-toydata2[,-4]
lambda<-data.frame(c(1:10),runif(10))
colnames(lambda)<-c('DocId', 'Lambda')
id<-seq(1:10)

iterative.bt(1,1,id,lambda,toydata2,20)

}
\author{
Benjamin Schneider, Zoe Ang, and Hyun Woo Lim
}
