% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRankPoints.R
\name{power_rank}
\alias{power_rank}
\title{Creates a Power Rank Points Prediction Model}
\usage{
power_rank(
  gameIds,
  homeTeamIds,
  awayTeamIds,
  homeScores,
  awayScores,
  isNeutralSite = FALSE
)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}
}
\value{
named list with team power ranks + home advantage values and functions to predict matchups based on team identifiers or provided strengths
}
\description{
This package creates a simple prediction model based on team identifiers (potentially names or ints) and previous game results.
}
\examples{

}
