% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOlsRating.R
\name{team_ols_rating}
\alias{team_ols_rating}
\title{Creates a Team OLS Optimized Rating Prediction Model}
\usage{
team_ols_rating(gameIds, homeTeamIds, awayTeamIds, homeScores, awayScores,
  isNeutralSite = FALSE)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}
}
\value{
named list with team and home field strength values and functions to predict matchups based on team identifiers or relative strengths
}
\description{
This package creates a simple prediction model based on team identifiers and previous game results.
}
\examples{

}
