% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreDeviation.R
\name{z_score_dev}
\alias{z_score_dev}
\title{Creates a Z-Score Deviation Prediction Model}
\usage{
z_score_dev(
  gameIds,
  homeTeamIds,
  awayTeamIds,
  homeScores,
  awayScores,
  isNeutralSite = FALSE
)
}
\arguments{
\item{gameIds}{}

\item{homeTeamIds}{}

\item{awayTeamIds}{}

\item{homeScores}{}

\item{awayScores}{}
}
\value{
named list with team strengths + home/away adjustment values and functions to predict matchups based on team identifiers or provided strengths
}
\description{
This package creates a simple prediction model based on team identifiers and previous game results.
}
\examples{

}
