% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradleyterry.R
\name{bradleyterry}
\alias{bradleyterry}
\alias{bradleyterry,}
\alias{ANY-method}
\title{Bradley-Terry}
\usage{
bradleyterry(a, b, lambdai, lambdaj, dataset)
}
\arguments{
\item{a}{an a value}

\item{b}{a b value}

\item{lambdai}{a vector of lambda_i}

\item{lambdaj}{a vector of lambda_i}

\item{dataset}{a dataframe with three columns of data}
}
\value{
an output lambda_i^(t)
}
\description{
Finds the probability of a student getting certain exam questions correctly given a parameter theta.
}
\note{
This function produces vectors for use in finding the likelihood function.
}
\examples{

lambdai<-runif(30)
lambdaj<-runif(30)
toydata<-data.frame(rep(1,30), sample(1:100, 30, replace=T), sample(c(0,1), 30, replace=T))
colnames(toydata)<-c("DocIDi", "DocIDj", "Choose")
bradleyterry(1,2,lambdai,lambdaj,toydata)

}
\author{
Benjamin Schneider, Zoe Ang, and Hyun Woo Lim
}
